Alteracoes na transacao, c.credito(atributo passou para long long int), multibanco

+++ CartaoCredito.h
@@ -10,8 +10,8 @@ private:
     Date validade;
 public:
     CartaoCredito();
-    CartaoCredito(long int numCartao, Date val);
-    CartaoCredito(double valor, long int numCartao, Date val);
+    CartaoCredito(long long int numCartao, Date val);
+    CartaoCredito(double valor, long long int numCartao, Date val);
     long int getNumCartao() const;
     void setNumCartao(long int numCartao);
     void setValidade(Date validade);
